Update docker-compose and main.go, and add .idea to .gitignore
In docker-compose.yml, the port numbers have been updated to strings to ensure consistent interpretation across different systems (from 8080:8080 to "8080:8080", from 80:80 to "80:80").

In backend/main.go, a double import of the godotenv package was detected and removed to avoid redundancy and potential errors during runtime.

A .idea rule was added to .gitignore to prevent JetBrains IDE settings from being tracked, as they contain user-specific settings and should not be included in the version control system.